libflux: use a stack for non-glob msg request handlers
Problem: When a duplicate request message handler for a non-glob topic
string is registered with the dispatcher, it replaces the exisiting
message handler in the dispatcher "handlers_method" hash. When either
the original or new message handler is destroyed the hash entry is
purged and therefore messages are dropped for what should be an active
message handler.

Introduce a stack structure which replaces the single message handler
stored in method handlers hash. This allows new request handlers that
match existing handler to be pushed onto the stack without losing track
of existing  handlers. When existing message handlers are destroyed,
they are simply removed from the stack. If the new message handler
is destroyed, then it is popped from the stack and the old, active
message handler takes over again.

Fixes #3476