job-manager: fix possible leak on jobtap plugin load
Problem: When a jobtap plugin is loaded, currently active jobs are
iterated so that the plugin's "job.new" callback can be called
for each active job. When creating the list of active jobs from
the `ctx->active_jobs` hash, however, the current code removes the
default zlistx destructor in the erroneous assumption that there was
no corresponding default destructor. This results in a leak of all
jobs that were active at the time of a jobtap plugin load.

Remove the unnecessary call to

    zlistx_set_destructor (jobs, NULL);

from `jobtap_handle_load_req()` to prevent the leak.