job-manager: jobtap: remove workaround in plugin load
Problem: A workaround exists in jobtap plugin load to prevent
destruction of a matching handler in the message dispatcher, but
the underlying problem has been fixed, so the workaround in no
longer necessary.

Remove the code in jobtap_load() which unloads any currently loaded
plugin before loading a new plugin. This will not work with multiple
plugins, and is no longer necessary.