python: update asyncio integration
The clever implementation of running the Flux event loop only
when there are updates to process can hang the asyncio event
loop when there are large numbers of Flux events to process.
A dumber implementation, of running the Flux event loop for
one iteration every asyncio loop iteration, does not have this
problem, although it may waste cycles when there are no events
to process.