testsuite: use bulksubmit for inception tests
Problem: The CI "inception" tests use prove(1) to run each sharness
test as a Flux job, but prove is meant for console use and is thus
unwieldy in CI.

Add a script, t/test-inception.sh which runs the tests using
`flux mini bulksubmit` instead of prove(1). Dump the output of any
failing tests to <test>.log so that the checks-annotate.sh script
can find output. Exit with failure if any test failed.