mergify: dismiss reviews on updated PRs from non-maintainers
Problem: When a pull request has been approved and had the
merge-when-passing label set, new, unreviewed code could be pushed
to the pull request and the unreviewed code will be merged.

Add a rule to the mergify config to dismiss approving reviews whenever
a pull request is updated from a contributor who is not on the core
team.