python: do not create empty headers in make_clean_header.py
Problem: When the python binding utility script `make_clean_header.py`
fails it leaves an empty header file in the build directory. If make
is run again, then this header file is assumed to be a product of
a previously successful build, and the rest of the project build
continues, only to fail later in the testsuite when python modules
aren't built correctly or at all.

Move the open of the target header file in `make_clean_header.py` to
the end of the script, when it is guaranteed that the header will be
successfully created, in order to avoid the empty header file problem.