kvs-watch: add cancel/disconnect credential check
Problem: kvs-watch is open to guest users, but the cancellations
and disconnects of the watches do not verify that the request came from the
same guest user.

Solution: When matching messages, call new convenience functions
flux_disconnect_match() or flux_cancel_match() to ensure the credential
of the requestor is valid in addition to the route/matchtag matching.
The new convenience functions handle matchtag matching, therefore
there is no longer a need to parse the matchtag.

Fixes #3528