libccan: import CCAN list implementation
Problem: the zlist[x] container internally allocates a list
node for each list item.  When a list has high traffic, the
heap activity and the disjoint memory addresses of node and item
could have a higher overhead than a list implementation that
can embed the node in the item.

Add the CCAN list implementation, which is similar to the linux
kernel's, and embeds the list node in the item.

Source taken from:
  https://github.com/rustyrussell/ccan.git
  commit 48b4ffc38ecef0d423a09967e6784adb5807033a

The libccan tree was created by running CCAN's create-ccan-tree
utility, which resolves dependencies, resulting in a couple of
other CCAN modules being pulled in.  The Makefile.am was then
modified to embed with Flux.

The license is BSD-MIT and CC0, both compatible with our LGPL.

Omit the foreign code from code coverage analysis.

Note: one of the dependencies creates a streq() macro that conflicts
with one of the same name (and equivalent implementation) in czmq.h,
so that has to be #undefed when both are used.

Add -I$(top_srcdir)/src/common/libccan to AM_CPPFLAGS where CCAN is
used internally.