content-cache: avoid scanning for dirties on flush
Problem: on rank 0, the RPCs to backing store are batched to
avoid having too many outstanding, but once more stores can be
sent, the dirty cache entries have to be re-discovered by
performing a linear search on the hash.

Create a list for dirty entries that have not yet been stored,
so that the linear search can be avoided.