content-sqlite: ensure coverage of batch store
Problem: codecov report shows that test to supposedly exercise
the content-cache store batching on rank 0 when a backing store
is loaded is not actually exercising that code.

Probably the synchronous stores from the test take about as long
as the stores from the cache to sqlite, so no backlog is created
by the test.

Solution: overlap the content store RPCs from the test.