kvs: send error message on namespace corner case
Problem: Users will receive an EEXIST errno when a namespace is being
created, but a namespace of that name already exists.  They will receive
this errno regardless if the namespace is active or in the process of being
removed.  If it is in the process of being removed, it can be confusing
why an EEXIST error was received.

Solution: If a namespace is in the process of being removed, return an
error message giving details of why EEXIST was returned.

Fixes #3637