content-cache: leave min sync period unbounded
Problem: cache entries are purged every heartbeat with the period
bounded by sync_min=1 and sync_max=10 seconds.  In test we would
like to crank down the heartbeat period to make the test run faster
but setting it less than sync_min doesn't help.

sync_min was established to avoid triggering the heavyweight cache
purge too frequently, for example when heartbeat messages "bunch up"
in the message queue.  Now that purging uses an LRU, it doesn't need
to scan for eligible items, and exits immediately if there is nothing to do.

Eliminate the sync_min lower bound on the cache purge period.