flux-start: add --test-hosts=HOSTLIST option
Problem: flux-start does not have support for starting brokers
that bootstrap from a config file, but that would be very handy
for running tests that involve restarting brokers, since PMI
is not restartable.

Add --test-hosts=HOSTLIST option, which sets FLUX_FAKE_HOSTNAME
for each broker to the Nth element of the hostlist, where N is
the broker rank.

Add coverage for the new option to t0001-basic.t.