testsuite: build shell plugins with fluxplugin_ldflags
Problem: Shell and jobtap plugins built in the testsuite set a bad
example and do not all use the correct LDFLAGS. Notably, -avoid-version
is missing so plugins are unnecessarily built with shared library
versioning.  Other projects may copy the Makefile as an example,
so these plugins should be built properly.

Use $(fluxplugin_ldflags) when building all test plugins in the
testsuite, thereby setting a proper example.