flux-start: add embedded server for testing
Problem: tests that involve starting and stopping brokers are
difficult to orchestrate using flux-start, but we will need
support for running such tests in CI.

Use the new usock_service to embed a server in flux-start.
The server creates a socket named 'start' in the rundir, and
sets FLUX_START_URI in the environment for clients.

Currently the server has support for the following methods:

start.status
  Return an array of procs that includes broker PIDs in rank order

disconnect
  Log receipt of disconnect message.  This is a placeholder for
  future streaming socket management.

Add a test front end command that provides the client side for the
start.status RPC, and is available to add sub-commands for simple,
shell script driven testing.  More sophisticated, event driven test
programs would be written in python and combine broker and flux-start
communication.

Add a few tests to t0001-basic.t to exercise basic function.