kvs: maintain queue of KVS roots with ready txns
Problem: The transaction prep and check watchers iterate
through all KVS roots looking for / processing KVS transactions.
If there are many KVS namespaces, this iteration can take up a
significant amount of processing time.  This processing time can be
wasteful if the majority of the namespaces do not have any
processing to do.

Solution: Maintain a "work queue" of KVS roots that need to
be processed, thereby limiting the amount of iteration that is
done.

Fixes #3633