flux-proxy: check broker version
Problem: problems often arise when flux-proxy from one version
of flux is used to connect to another version of Flux.

Add a version check so that flux-proxy does not start up if
the broker version does not match that of flux-proxy.

A --force option is added to bypass this check with only a warning.

Update flux-proxy(1) manual page.

Fixes part of #1816