job-manager: cache current job for jobtap plugins
Problem: When a jobtap plugin calls back into the job-manager, the
current job on which the call has been made has be looked up by id
in the active jobs hash. This is unnecessary, and additionally means
that the `struct job *` can't be used in jobtap calls before the
job has become active, e.g. "job.new" and "job.validate".

Since jobtap only operates on a single job at once, cache the current
job in the jobtap structure so that there is quick access to this
object, as well as possible access when the job is not yet active.