job-manager: jobtap: allow dependencies before submit event
Problem: It would be convenient if dependency-add and remove events
were supported before the "submit" event is posted. This would simplify
plugins that would like to validate and generate dependencies in a
single callback. This would also prevent plugins from unnecessarily
parsing jobspec multiple times.

When a job has dependencies added before it is technically "active"
(i.e. before the submit event, in the job.validate and job.new
callbacks), add these dependency events to an ephemeral list attached
to the job's aux_item list. Upon entry to the DEPEND state, these
dependencies will all be emitted to the eventlog as if they were done
in the job.state.depend callback.

This allows jobtap plugins to emit dependencies in job.validate. If
the job is subsequently rejected, then the dependencies are a no-op.
If the dependencies fail to be posted in the DEPEND state, then a
job exception is raised.