testsuite: change name of jobspec attr in dependency tests
Problem: attributes.system.dependencies is a reserved attribute
in canonical jobspec, but the t2270-job-dependencies.t test uses
it to drive testing of dependencies.

Change the name of the attribute used by the dependency-test.so
plugin and t2270-job-dependencies.t to avoid conflict with the
real dependency array in jobspec.