job-manager: process dependencies during submit validation
Problem: The job-manager is currently unaware of the
attributes.system.dependencies list for jobs, and while a jobtap plugin
can process this list and generate dependency-add and remove events,
there is no way to determine if all listed dependency objects in the
jobspec were handled.

Add a function to jobtap to iterate jobspec dependencies and issue a
`job.dependency.<scheme>` callback for each dependency. Jobtap plugins
can then add a handler for scheme that they support. If no handler
is registered for a given scheme, then the job is rejected with a
descriptive error.

Call `jobtap_check_dependencies()` during submit handling just after
other job validations. This ensures that any error message is returned
as a submission error, instead of a later job exception. This also
means that jobtap plugins which support dependencies can do validation
in their job.dependency.* callback, instead of requiring a separate
validation function.

Fixes #3651