flux-mini: support --dependency option
Problem: There is no way to specify RFC 26 "shorthand" dependencies
on the flux-mini commandline and it is inconvenient to specify a
set of dependencies with --setattr since they are an array of
dictionaries.

Add a `--dependency` option which takes RFC 26 shorthand URIs on
the commandline, e.g.

 --dependency=fluid:f1234?type=in

will be expanded to:

 { "scheme": "fluid", "value": "f1234", "type": "in" }

Multiple --dependency options may be used to append multiple objects
to the attributes.system.dependencies array in jobspec.