flux-mini: do not expand flux submit args without --cc/bcc
Problem: "Mutable" option values are substituted in flux mini submit
in case `--cc` or `--bcc` are used to allow substitution {cc} to be
used. However, this is inconvenient when --cc is not used since it
makes passing Python dictionaries impossible, since a substitution
attempt will be made on matching '{' and '}', and the escaped brackets
"{{" and "}}" are assumed to be mustache templates.

Since option value replacement without --cc doesn't do anything anyway,
disable it for `flux mini submit`, which will at least allow Python
dictionaries to be passed on the commandline in this one case.

For passing Python dicts to jobs submitted in bulk, the bulksubmit
command should instead be used and the Python objects passed in the
inputs (either on stdin or separated via :::).