flux-mini: improve substitution string error messages
Problem: The error messages when option value substution raises
an IndexError or KeyError may not be clear. On IndexError, a generic
"Invalid replacement string" error is emitted, and on KeyError, the
command just emits:

 flux-mini: ERROR: 'key'

Improve the error messages so that an invalid index in the substitution
notes that fact, and an invalid key specifies that the key was not found
and the name of the missing key.