kvs: use json dict over array in setroot event
Problem: In KVS setroot events, an array of keys is sent with the event
for listeners to determine what keys may have changed.  This has several
performance issues:

1) Profiling discovered that the normalize_and_append_unique()
function can be a bottleneck.  In the function, the array of keys is
regularly searched to determine if any keys are duplicates of each other.
This is an O(N^2) search.

2) Any listeners of the setroot event have to scan the entire array to
determine if a key they are interested in exists in the array.

Solution:

Replace the array with a json dict.  The dict will be much faster for
discovering duplicate keys.  On the listener side, lookups of keys will
be much faster.