flux-mini: allow --setattr to read JSON value from file
Problem: It is unwieldy to try to set even moderately complex
jobspec attributes on the --setattr commandline. This becomes
impossible when using --cc/--bcc and/or bulksubmit since this
option allows substitution with `{}`.

Allow the value of an attribute to be read from a JSON file when
the ATTR in the --setattr option begins with the "^" character.
This allows arbitrary attributes to be set easily in jobspec on
the flux mini run/submit/bulksubmit etc command line.