shell/openmpi.lua: set OMPI_MCA_btl_vader_backing_directory
Problem: shared memory segment naming for the "vader" BTL may collide
when running multiple jobs per node.

This is because the path is constructed from
  /dev/shm/vader_segment.hostname.uid.jobid.rank

and the flux pmix plugin in openmpi always sets the jobid to 0
because FLUX_JOB_ID in f58 format is converted to a 32-bit integer
using strtol().  This is a bug in the plugin, but a proper fix may
be elusive given the integer size mismatch of IDs.

Work around the plugin bug by changing
  OMPI_MCA_btl_vader_backing_directory
from default of /dev/shm to $FLUX_JOB_TMPDIR, which the shell tmpdir
plugin will have set to a job-unique directory.

Fixes #3649