kvs: limit iteration in msg disconnect path
Problem: Profiling showed that there was a small bottleneck in
the KVS disconnect path.  Most notably, the bottleneck was in the
call to cache_wait_destroy_msg(), as every entry in the cache
hash was iterated and every waitlist in every entry was iterated over.

Solution: Create two maintenance lists in the cache, "notdirty_list"
and "valid_list", to maintain lists that contain waitlist messages.
By maintaining these lists, we can limit the the iteration in
cache_wait_destroy_msg() to the bare minimum.