kvs: reduce overhead in cache expiration path
Problem: During profiling runs the cache_expire_entries() function
is shown to be a minor bottleneck.  It appears to be busy due to a
call to zhashx_keys().  The function creates a zlistx of all keys
in the primary cache's hash.  This requires iterating through the
entire hash (all hash buckets, including empty ones), and allocating
memory for every string added to the zlistx.

Solution: Support a list 'entries_list' which maintains a list of
all current cache entries.  When cache_expire_entries() is called,
simply iterate through this list instead.  It eliminates the iteration
of the entire cache hash and eliminates the excess memory allocations.