job-manager: change jobtap RPC protocol
Problem: The current jobtap RPC protocol is short-sighted in that
it assumes only a single plugin can be loaded at a time, so a load
request removes any currently loaded plugin. This will not be ideal
when a list of plugins are supported.

Add a `remove` member to the job-manager.jobtap RPC which contains a
plugin name or glob of loaded plugins to remove. As a convenience,
"all" is a synonym for "*" to easily allow all loaded plugins to
be removed.

The remove pattern can be sent along with a load request, in which
case all matching plugins are removed before the new plugin is loaded.

Add a `flux jobtap remove` command, as well as a new `--remove`
option to `flux jobtap load`. This allows the previous load behavior
to be preserved with either

  flux jobtap remove all
  flux jobtap load plugin.so

or

  flux jobtap load --remove=all plugin.so

The current plugin is no longer auto-unloaded when loading a new
plugin. Therefore, it is an error if a plugin is loaded while another
plugin is still loaded without first removing it via `flux jobtap
remove` or `flux jobtap load --remove=*`.

Update all tests which used `flux jobtap load`.