shell: add tmpdir plugin for FLUX_JOB_TMPDIR
Problem: MPI shell plugins and perhaps other future applications can
benefit from the availability of job-specific temporary directories.

Create a job-specific directory when the shell starts.
Set FLUX_JOB_TMPDIR to the new directory in the job environment.
In the shell's atexit handler, remove the directory recursively.
In case a job spans multiple brokers on the same node, avoid cleanup
races by including the shell rank in the directory name.

Attempt to create the directory in the broker's rundir.  If that
fails, as it would for guest users, fall back to ${TMPDIR:-/tmp}.

Note that cleanup by the shell is not reliable since it may be
terminated early.  However, that is mitigated in most cases by the
fact that the broker rundir is cleaned up by the broker itself or
by flux-start.  The system instance should employ an epilog script
or similar to clean up directories in ${TMPDIR:-/tmp}.

If TMPDIR was set in the job environment, ensure that it exists,
replacing the tmpdir.lua shell plugin, which was added to address #3326.

Since other plugins (such as the batch plugin) may depend on TMPDIR
and/or FLUX_JOB_TMPDIR, place this plugin at the top of the stack.