job-manager: jobtap: support plugin searchpath
Problem: Loading jobtap plugins requires a fully qualified path (or
path relative to the brokers cwd). This will be inconvenient when
loading plugins from configuration.

Add support for a plugin searchpath. If searchpath is set and a
relative or absolute path isn't specified for a plugin then the first
plugin found in the colon-separated search path is loaded.