libflux: module: ignore DSOs with no mod_name symbol
Problem: flux_modname(3) returns a dlerror() based error message
in any provided error callback when dlsym ("mod_name") fails, but
this extra error is probably unecessary in most, if not all,
situations. It also generates a lot of noise when flux_modfind(3)
recurses through directories with a lot of non-module .so files.

Simplify flux_modname(3) to return EINVAL when a "mod_name" symbol
is not found in the DSO. This will quiet unnecessary errors from
flux_modfind() and `flux module load` etc, effectively ignoring
DSOs without a `mod_name` symbol.