shell: fix flux_shell_setenvf() overwrite
Problem: flux_shell_setenvf() fails with EEXIST if a name
exists in the environment but overwrite=0.  This is different
from setenv(3) which returns success in that case, and thus
may be unexpected.  It also differs from its man page which
describes setenv(3) overwrite semantics.

Change flux_shell_setenvf() to return success if overwrite=0
and the name exists in the environment.