libsubprocess: fix flux_cmd_setenvf() overwrite
Problem: flux_cmd_setenvf() fails with EEXIST if a name
exists in the environment but overwrite=0.  This is different
than setenv(3) which returns success in that case, and thus
may be unexpected.

Change flux_cmd_setenvf() to return success if overwrite=0
and the name exists in the environment.

Update unit test.