build: add -avoid-version to job-manager LDFLAGS
Problem: The job-manager can't be built with $(fluxmod_ldflags)
since it needs to export `flux_jobtap_*` symbols as well the mod_*
symbols. However, when the Makefile was switched to $(fluxlib_ldflags)
the use of `-avoid-version` was dropped, and so job-mangaer.so
is installed versioned.

Add `-avoid-version` to the job-manager LDFLAGS to avoid building
a versioned DSO for this broker module.