job-manager: remove builtin default and hold jobtap plugins
Problem: The builtin.priority.{default,hold} jobtap plugins are not
useful. The default plugin is only meant for demonstration purposes
and behaves the same as if no plugin was loaded, and the "hold" plugin
will rarely be used so it is wasteful to compile it into the job-manager.

Move the builtin.priority.default plugin to a test-only plugin under
t/job-manager/plugins/priority-default.c.

Compile the builtin.priority.hold plugin as a separate DSO, installed
into a the default jobtap search path so it can be loaded with

 flux jobtap load priority-hold.so

Finally, do not load any jobtap plugins by default.

Update sharness tests as necessary.