job-manager: jobtap: drop requirement for flux_plugin_name
Problem: Requiring jobtap plugins to set a name and the using that
name in the plugin listing and removal tools is a bit arbitrary.

To make these interfaces simpler, use the basename of the plugin
DSO as the "name" reported in `flux jobtap list`, as well as the
argument to remove. This means a plugin can easily be reloaded as

 flux jobtap load --remove=plugin.so plugin.so

or

 flux jobtap load --remove=plugin.so /path/to/plugin.so

Plugins may still register a name which will be used to determine
the topic string for flux_jobtap_service_register(3). If no plugin
name is registered then the DSO name is used without the `.so`.