job-manager: jobtap: prevent loading plugin with duplicate name
Problem: The same plugin can be loaded multiple times in the job-manager
with `flux jobtap load` or via configuration. This could cause confusing
results.

Prevent loading the same plugin twice by refusing to load a plugin with
the same name as an already active plugin. A plugin can easily be
reloaded with

 flux jobtap load --remove=test.so test.so