shell/pmi: make "pershell" the default mapping
Problem: mvapich2 fails when PMI_process_mapping is not provided
by the PMI KVS.

The shell pmi plugin defaults to "none" if the process (broker)
mapping of the Flux instance cannot be determined.  This may happen,
for example, if run by (slurm) srun --mpi=none.

The default of "none" was chosen because it prevented openmpi shared memory
paths from colliding when there were multiple brokers/shells per node,
as it seems openmpi assumes it owns the node if there is no clique.
Now that the openmpi.lua shell plugin sets the shared memory path to
be within FLUX_JOB_TMPDIR, this is no longer a problem.

Change the default mapping to "pershell" to appease mvapich2.

Fixes #3592 (again).