job-manager: jobtap: do not match builtins with remove "*"
Problem: When removing jobtap plugins via configuration or with `flux
jobtap remove`, it will be convenient if a glob of "*" doesn't match
any builtin plugins.

Use the FNM_PERIOD flag with fnmatch(3) to duplicate behavior of the
shell and require matching names with a leading dot to explicitly
include a dot as the first character of the glob.

Builtin plugins may then start with a `.` character, simulating
hidden files.