job-manager: recheck dependencies on job-manager restart
Problem: Handling job-manager restart presents a tricky situation
for jobtap plugins which use job.dependency.<scheme> style callback
topics to subscribe to the dependency schemes handled by the plugin.
On restart, the job.dependency.* calls are not re-issued, so these
plugins would have to subscribe to job.new or job.state.depend
callbacks, parse the jobspec dependency section themselves to see if
there is a valid dependency for the scheme(s) of interest -- that is,
everything that the job.dependency.* callbacks are supposed to prevent.

To avoid this extra work for dependency plugins, add an additional
call to jobtap_check_dependencies() for each job in the DEPEND or any
previous state during job-manager restart. This allows a dependency
plugin to use the same callback handler for normal job submission
and restart.

Since multiple dependency-add events with the same description are
ignored, the plugin does not need to be aware that the job.dependnecy
calls are being made under restart. This allows the plugin to set up
any necessary internal state in the same way for restart or normal
job submission, reducing code complexity and duplication for plugin
authors.