job-manager: jobtap: call any dependency callbacks on plugin load
Problem: When jobtap plugins are loaded any job.dependency.*
callbacks are not issued for jobs in the DEPEND state. This could
cause jobs to get stuck in DEPEND state if a dependency plugin
or job-manager is reloaded.

Refactor jobtap_check_dependency() to allow optionally calling the
job.dependency.* callback on a single plugin instead of the whole
stack.

Use this function to add a new plugin_check_dependencies() call which
may be made from jobtap_load_plugin() to call any dependency callbacks
in the newly loaded plugin for all active jobs in DEPEND state
at plugin load time.