job-manager: add builtin begin-time dependency plugin
Problem: A minimum begin or "start-after" time for jobs is not
supported by Flux.

Add a simple begin-time dependency plugin to the job-manager.  If a
dependency with scheme "begin-time" exists in jobspec, then the value
is assumed to be floating point seconds since epoch at which the job
should be made eligible to run.