testsuite: enhance job depedency plugin tests
Problem: Support for calling dependency plugin callbacks during
job-manager restart and plugin load are not tested in the testsuite.

Enhance the dpeendency-test.so plugin with support for creating
fake plugin "state", along with a service to query if the state is
currently active for a given job & dependency.

Ensure plugin job.dependnecy.* callbacks are called on both job-manager
resart and plugin load in t2270-job-dependencies.t.