job-manager: add requisite builtin default priority plugin
Problem: The job-manager requires special code in the jobtap module
to handle the case where there is no current priority plugin loaded,
and takes default action of setting priority to urgency in this case.
Now that multiple plugins are supported, it should be possible to
instead have a default priority plugin as a builtin, thus encapsulating
the "default" behavior.

Move the default priority plugin back from a test-only plugin to a
builtin plugin. Remove the code in the job-manager which handled the
special case of no priority plugin loaded. Now if the .priority-default
plugin is ever unloaded, jobs will stick in PRIOIRTY state until a
plugin is loaded which does handle priority setting for jobs.