libflux: make FLUX_PLUGIN_ARG_UPDATE the default
Problem: In the most common use case, FLUX_PLUGIN_ARG_UPDATE is the
desired behavior for packing and setting both in and out plugin args,
but this is not the default behavior.

Switch the default behavior of flux_plugin_arg_set/pack to update
plugin args instead of replacing. Add a FLUX_PLUGIN_ARG_REPLACE
flag to restore the old behavior and replace args.

Update all internal callers.

Fixes #3683