job-manager: jobtap: raise job exception on priority unpack failure
Problem: If there is a failure to unpack job priority after calling
all jobtap plugin job.state.priority callbacks, then the job may end
up mysteriously stuck in the PRIORITY state. This condition could be
frustrating for users and difficult to diagnose for admins.

Generate a non-fatal job exception when this occurs, with a note that
the job may be stuck in the PRIORITY state until the plugin is fixed
and reloaded.