doc: update jobtap documentation
Problem: flux-jobtap(1) and flux-jobtap-plugins(7) are out of date
and missing documentation of new features.

Update the flux-jobtap-plugins(7) manpage with information about new
handling of missing priority during PRIORITY state transition, as
well as addition of new sections for plugin naming and priority
handling.

Update the flux-jobtap(1) manpage to mention the use of the -a, --all
option to `flux jobtap list` to include builtin plugins in the listing.