testsuite: set exit-on-error in MPI_Abort test
Problem: mvapich and some mpich versions are explicitly not tested
in the MPI_abort sharness test because they perform an early task
exit rather than sending a PMI-1 abort command.  Now that we have
early task exit detection, we can test those versions.

Drop the 'skip' logic from the test, set -o exit-on-error, and relax
the grep for the exception in the output message to allow for variation.

Add a note documenting the MPI behaviors that were previously
embodied in the skip logic so we don't lose that information.

Fixes #3690