shell: add exit-on-error option
Problem: it may be useful for a parallel job to abort immediately
when any task fails, as opposed to waiting for the exit-timeout to
expire.

Add an option 'exit-on-error' that causes shell-0 to call shell_die()
with the failing task's exit code.

Also:  change the 'exit-timer' handler to print the failing rank and
exit with the failing task's exit code rather than 1.

Add test coverage for exit-on-error.

Fixes #3691