testsuite: add "system" sharness test personality
Problem: brokers cannot be restarted in tests if they are
bootstrapped with PMI.

Add a "system" personality to test_under_flux, in which a [bootstrap]
config and curve certificate are generated and stored in the test's trash
directory for the requested size instance.

Since sharness prophylactically removes and re-creates the trash directory
before launching the test script, re-execute the script with --root set
to the original trash directory path, so config and curve certificate
are not removed.