job-manager: always capture end_event in struct job
Problem: The "end event" (the event that causes transition to CLEANUP)
is only captured for waitable jobs, but it may be convenient to have that
event available for plugins.

Capture the end_event for all jobs. This will cause a small amount of
extra memory use for jobs after the CLEANUP state, but this should be
short-lived for non-waitable jobs anyway.