job-manager: jobtap: add flux_jobtap_get_job_result()
Problem: There is not a convenient way for a jobtap plugin to get the
"result" of a job, i.e. whether it succeeded, failed, or was canceled.

Add a convenience call `flux_jobtap_get_job_result()` that
allows a caller to get the `flux_job_result_t` of a job from the
job.state.cleanup or job.state.inactive callbacks.