job-manager: post dependency-remove events after all add events
Problem: A list of pending dependency events are added to an internal
dependencies list of jobs when a jobtap plugin emits dependencies while
a job is in the NEW state (e.g. job.new and job.validate callbacks.) If
a plugin emits an add and remove event in this context, then later this
may result in a job being prematurely moved past the DEPEND state if
the add and remove are allowed to occur in succession. This happens
even though other plugins may also have pending dependency-add events.

Process all pending dependency-add events before dependency-remove
events, this ensures that the set of dependencies is fully populated
before references start being cleared by dependency-remove events.