job-manager: add simple jobid dependency plugin
Problem: Flux does not support simple active job dependencies.

Add a builtin job-manager plugin for handling simple jobid dependencies.

The following URIs are currently supported:

 after:ID
    dependency is removed after ID has entered RUN state

 afterany:ID
    dependency is removed after ID has entered INACTIVE state

 afterok:ID
    dependency is removed after ID has entered INAACTIVE state and
    its result is FLUX_JOB_RESULT_COMPLETED (i.e. a `finish` event
    with status=0 caused the transition to CLEANUP).

 afternotok:ID
    dependency is removed after ID has entered INACTIVE state and
    its result is not FLUX_JOB_RESULT_COMPLETED (i.e. the job
    finished with status != 0 or an exception caused the transition
    to CLEANUP state)

The plugin works by installing a list of dependent job and dependency
type on jobs that are the target of the dependency (so called requisite
jobs). When a job enters the RUN and INACTIVE states, presence of this
list is checked, and if it exists, dependencies of the appropriate
type are immediately satisfied.

When a dependency is added to another job, a reference list is also
added to the dependent job. This allows the dependent job to remove
itself from any dependency lists under abnormal termination, e.g.
if the job is canceled while in the DEPEND state.

After a requisite job enters the INACTIVE state and all satisfied
dependencies are handled, any remaining dependent jobs -- i.e. those
jobs with dependencies that can now never be satisfied -- are terminated
with a fatal "dependency" exception.

Users can only place dependencies on their own jobs.