flux-start: add --test-exit-mode=leader
Problem: testing that restarts brokers requires a cleanup
mode that does not start the self-struct sequence when a broker
stops, because the test might want to restart it and continue on.

Add the option --test-exit-mode with possible values of "any",
which means any broker exiting starts the exit timer, as before;
or "leader", a new mode that only starts the exit timer when the
leader broker exits.

Also, flux-start's exit code is handled as follows:
leader - exit with broker 0's exit code
any - exit with the higest exit code of all brokers