flux-start: handle error execing broker
Problem: if broker cannot be exec()ed, flux-start does not
print an error message.

Call log_err_exit() if the exec fails, and ensure that errno
is preserved on that call path.

Also, it appears from testing that an exec error is reported
immediately by flux_exec(), not in the state callback as
FLUX_SUBPROCESS_FAILED.  In fact looking at libsubprocess, I'm
not sure that state is ever entered.  Change its handler to
just exit immediately with an error.