flux-sharness: add opt to start only broker rank 0
Problem: sharness tests running under the 'system' test
personality need to be able to configure the flux start
--test-start-mode=MODE option.

Add an environment variable TEST_UNDER_FLUX_START_MODE
that can be set to 'leader' before test_under_flux is called
to have the test script with only the rank 0 broker started.
The test script would be expected to bring up other brokers
as needed.

Generate a fake 'R' for the system test personality, and set
the noverify flag.  This enables the resource module and scheduler
to service requests with only the rank 0 broker online.