optparse: optparse_get_int: return option use count if no arg
Problem: optparse_get_int() generates a fatal error if it used with
an option that doesn't take an argument. However, it would be useful
to use this function when the argument is optional, by returning the
number of times the option was used.

If no option argument was found with optparse_get_int(), then return
the number of times option was used. Update header documentation and
tests.