optparse: make optional arg apply to longopt only by default
Problem: It is common practice to make a `-v, --verbose` option
increase verbosity with multiple use, e.g. `-vvv` sets `verbose = 3`
internally in the command or process to which the option is applied.
If using the long option, however, it would be more convenient if an
optional argument was used, so that `--verbose=3` is equivalent to
`-vvv` (Nobody wants to write `--verbose --verbose --verbose). Sadly,
this is not currently supported in an acceptable way by liboptparse. If
you choose an optional argument, then `-v` cannot be combined with
any other short option, since `-vv` passes `v` as the argument to
the first `-v`.

Change liboptparse so that optional argument (.has_arg = 2) applies
only to the long option by default. The old behavior can be
restored with the new flag: OPTPARSE_OPT_SHORTOPT_OPTIONAL_ARG.